refactor(cpp): consolidate native runtime tree
