+++ cpp/modalities/src/action_cpu.cpp
@@ -1,4 +1,4 @@
-#include "flashrt/modalities/action.h"
+#include "flashrt/cpp/modalities/action.h"
 
 #include <algorithm>
