+++ cpp/modalities/src/types.cpp
@@ -1,4 +1,4 @@
-#include "flashrt/modalities/types.h"
+#include "flashrt/cpp/modalities/types.h"
 
 #include <algorithm>
 #include <cstring>
