+++ cpp/modalities/src/vision_cpu.cpp
@@ -1,4 +1,4 @@
-#include "flashrt/modalities/vision.h"
+#include "flashrt/cpp/modalities/vision.h"
 
 #include <algorithm>
 #include <cmath>
