+++ cpp/models/pi05/src/spec.cpp
@@ -1,4 +1,4 @@
-#include "flashrt/models/pi05/cpp_runtime_spec.h"
+#include "flashrt/cpp/models/pi05/spec.h"
 
 #include <algorithm>
